Did some basic resource and heap copying
* Also changed size_t to UINT for framecount to remove type conversion warnings

+++ Support/Window.cpp
@@ -245,7 +245,7 @@ void DXWindow::EndFrame(ID3D12GraphicsCommandList10* cmdList)
 
 bool DXWindow::GetBuffers()
 {
-    for (size_t i = 0; i < FrameCount; ++i)
+    for (UINT i = 0; i < FrameCount; ++i)
     {
         if (FAILED(m_swapChain->GetBuffer(i, IID_PPV_ARGS(&m_buffers[i]))))
         {
