+++ Support/Window.h
@@ -32,8 +32,8 @@ public:
 		return m_isFullscreen;
 	}
 
-	static constexpr size_t FrameCount = 2;
-	static constexpr size_t GetFrameCount()
+	static constexpr UINT FrameCount = 2;
+	static constexpr UINT GetFrameCount()
 	{
 		return FrameCount;
 	}
